extract NoSearchProduct.jsx
